+++ view.php
@@ -34,7 +34,7 @@ require_login($course, true, $cm);
 $context = context_module::instance($cm->id);
 require_capability('mod/vimigallery:view', $context);
 
-$event = \core\event\course_module_viewed::create([
+$event = \mod_vimigallery\event\course_module_viewed::create([
     'objectid' => $gallery->id,
     'context' => $context,
 ]);